Move keepalive tests to a separate test file. (#3014)
This is a follow-up to https://github.com/grpc/grpc-go/pull/2992.
